add handle exception

+++ kcs-backend/src/main/java/group7/se1876/kcs_backend/repository/ProductRepository.java
@@ -3,5 +3,5 @@ package group7.se1876.kcs_backend.repository;
 import group7.se1876.kcs_backend.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 }
